SOA/order-service/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
motor==3.3.2
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.1
httpx==0.26.0
jinja2==3.1.2 